[Fix]: adding condition to stop recursion Modified: 3-O, 3-quick_sort.c

+++ 0x1B-sorting_algorithms/sort.h
@@ -23,7 +23,7 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-void _qsort(int *a, size_t low, size_t high, size_t size);
+void _qsort(int *a, int low, int high, int size);
 void shell_sort(int *array, size_t size);
 void _ssort(int *a, int size, int n);
 void cocktail_sort_list(listint_t **list);
